+++ execute-generators/src/main/kotlin/de/itemis/mps/gradle/generate/GenerateArgs.kt
@@ -18,7 +18,7 @@ class GenerateArgs(parser: ArgParser) : Args(parser) {
         .default(0)
         .addValidator { if (value < 0) throw InvalidArgumentException("parallel-generation-threads must be >= 0") }
     val disableStrictMode by parser.flagging(
-        "--disable-strict-mode",
+        "--no-strict-mode",
         help = "Disable strict generation mode. Strict mode places additional limitations on generators, but is required for parallel generation"
     ).addValidator {
         if (parallelGenerationThreads != 0 && value) throw InvalidArgumentException("strict mode is required for parallel generation")